Upload completed files from lab
Included is a python file that creates a Flask app, communicates inputs from the user to the Facebook Blenderbot model and returns responses, and records the chat history to give the chatbot context. Also included are the Docker file and it's requirements.txt., and an interface already provided by IBM as building the interface is not within the scope of this lab.